Consultas especificas y adición de relaciones en entidades

+++ src/main/java/gt/edu/umg/programacion2/proyectoFinal/sesionDeBienestar/repository/EmpleadoRepository.java
@@ -5,7 +5,6 @@
 package gt.edu.umg.programacion2.proyectoFinal.sesionDeBienestar.repository;
 
 import gt.edu.umg.programacion2.proyectoFinal.sesionDeBienestar.entity.Empleado;
-import java.math.BigInteger;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -15,6 +14,6 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface EmpleadoRepository extends JpaRepository<Empleado, BigInteger> {
+public interface EmpleadoRepository extends JpaRepository<Empleado, Long> {
     
 }
